26_commit
Refactor/Improved endpoints for Cart, Product, and Room
- Standardized response formats/type to JSON for consistency
- Implemented additional error handling mechanisms

+++ src/main/java/com/example/ApartmentRenovationCostEstimate/room/RoomRepository.java
@@ -4,5 +4,5 @@ package com.example.ApartmentRenovationCostEstimate.room;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface RoomRepository extends CrudRepository<Room, Integer> {
+public interface RoomRepository extends CrudRepository<Room, Long> {
 }
